pyproject.toml
[project]
name = "repl-nix-drivequest-builder"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.0.3",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "oauthlib>=3.2.2",
    "psycopg2-binary>=2.9.10",
    "google-api-python-client>=2.151.0",
    "requests",
]
